getRank and getSuit are implemented. log2 was written in an own method. annotations were updated
comment:
the way from the internet to calculate the log2 is bullshit.
i have implemented it and the result is, that this method is a stupid
way to do another very easy thing.
here the results if you implement the first part of the function:

def log2(pCard:Long):Long= {
    var x = pCard
    x |= (x >> 1);
    x |= (x >> 2);
    x |= (x >> 4);
    x |= (x >> 8);
    x |= (x >> 16);
    
    return x
  }

result:
pCard = 1 => 1
pCard = 2 => 3
pCard = 4 => 7
pCard = 8 => 15
pCard = 16 => 31
...
if you dont see it^^:
the first part is doing:
pCard = n => 2*n -1
XD

if you find a better way to calculate the logarithm, please do it :)
i could improve the function if i had the continue statement,
but it is not contained and i dont want to implement it.